// CareerDevelopmentHubCode/apps/career-development-hub/src/generated/models/follow-up-model.ts
import type { JobApplication } from './job-application-model';
import type { NetworkingContact } from './networking-contact-model';

export const FollowUpRelatedTypeKeyToLabel = {
  'Contact': 'Contact',
  'Application': 'Application',
  'NoneStandalone': 'None/Standalone'
} as const;
export type FollowUpRelatedTypeKey = keyof typeof FollowUpRelatedTypeKeyToLabel;

export const FollowUpReminderSnycStatusKeyToLabel = {
  'NotSynced': 'Not synced',
  'Synced': 'Synced',
  'Conflict': 'Conflict',
  'Error': 'Error'
} as const;
export type FollowUpReminderSnycStatusKey = keyof typeof FollowUpReminderSnycStatusKeyToLabel;

export const FollowUpStatusKeyToLabel = {
  'Open': 'Open',
  'Completed': 'Completed'
} as const;
export type FollowUpStatusKey = keyof typeof FollowUpStatusKeyToLabel;

export interface FollowUp {
  /**
   * @displayName Follow Up
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Title
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName Completed Date
   */
  completedDate?: string;
  /**
   * @displayName Due Date
   * @validationRule Required for create/update operations
   */
  dueDate: string;
  /**
   * @displayName Notes
   */
  notes?: string;
  /**
   * @displayName outlook_calendar_id
   */
  outlookCalendarId?: string;
  /**
   * @displayName outlook_event_id
   */
  outlookEventId?: string;
  /**
   * @displayName Related Application
   */
  relatedApplication?: Pick<JobApplication, 'id' | 'role'>;
  /**
   * @displayName Related Contact
   */
  relatedContact?: Pick<NetworkingContact, 'id' | 'contactName'>;
  /**
   * @displayName Related Type
   * @validationRule Required for create/update operations
   */
  relatedTypeKey: FollowUpRelatedTypeKey;
  /**
   * @displayName reminder_all_day
   */
  reminderAllDay?: boolean;
  /**
   * @displayName reminder_enabled
   */
  reminderEnabled?: boolean;
  /**
   * @displayName reminder_end_at
   */
  reminderEndAt?: string;
  /**
   * @displayName reminder_last_synced_at
   */
  reminderLastSyncedAt?: string;
  /**
   * @displayName reminder_sync_status
   */
  reminderSnycStatusKey?: FollowUpReminderSnycStatusKey;
  /**
   * @displayName reminder_start_at
   */
  reminderStartAt?: string;
  /**
   * @displayName reminder_sync_error
   */
  reminderSyncError?: string;
  /**
   * @displayName reminder_time_zone
   */
  reminderTimeZone?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: FollowUpStatusKey;
}

export const _FollowUp = 'FollowUp' as const;